pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ap-intake-validation"
version = "0.1.0"
description = "AP Intake & Validation pilot using Docling"
authors = [
    {name = "AP Team", email = "ap-team@company.com"},
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic>=2.7.4,<3.0.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "celery==5.3.4",
    "redis==5.0.1",
    "pika==1.3.2",
    "docling==2.60.1",
    "python-multipart==0.0.6",
    "aiofiles==23.2.1",
    "langgraph==1.0.2",
    "langchain==1.0.3",
    "langchain-core==1.0.3",
    "langsmith==0.4.41",
    "boto3==1.34.0",
    "minio==7.2.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dateutil==2.8.2",
    "pytz==2023.3",
    "email-validator>=2.2.0",
    "pandas==2.1.4",
    "sentry-sdk[fastapi]==1.40.0",
    "loguru==0.7.3",
    "langfuse==2.9.0",
    "prometheus-client==0.19.0",
    "jinja2==3.1.2",
    "lz4==4.3.2",
    "langgraph-checkpoint>=3.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "openai>=1.0.0",
    # Gmail API dependencies
    "google-api-python-client==2.108.0",
    "google-auth-httplib2==0.1.1",
    "google-auth-oauthlib==1.1.0",
    "google-auth==2.25.2",
    # QuickBooks dependencies
    "python-quickbooks==0.9.12",
    "intuit-oauth==1.2.6",
    "httpx==0.25.2",
    "reportlab>=4.4.4",
]

[dependency-groups]
dev = [
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "pre-commit==3.6.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-xdist==3.3.1",
    "pytest-mock==3.12.0",
    "pytest-benchmark==4.0.0",
    "httpx==0.25.2",
    "factory-boy==3.3.0",
    "faker==20.1.0",
    "psutil==5.9.6",
    "pytest-factoryboy==2.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers --strict-config --cov=app --cov-report=html --cov-report=term-missing"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "performance: marks tests as performance tests",
    "e2e: marks tests as end-to-end tests",
    "load: marks tests as load testing",
]
asyncio_mode = "auto"
